refactored: all components from login module moved to ui module
