+++ core/design-system/src/main/res/drawable/ic_launcher.png
