+++ core/design-system/src/main/res/drawable/ic_launcher_monochrome.png
