+++ core/design-system/src/main/res/mipmap-hdpi/ic_launcher.png
