+++ core/design-system/src/main/res/mipmap-mdpi/ic_launcher_foreground.png
