+++ core/design-system/src/main/res/mipmap-xhdpi/ic_launcher_monochrome.png
