+++ core/design-system/src/main/res/mipmap-xxhdpi/ic_launcher.png
