+++ core/design-system/src/main/res/mipmap-xxxhdpi/ic_launcher_background.png
