+++ core/ui/build.gradle.kts
@@ -24,4 +24,5 @@ android.namespace = "dev.enesky.core.ui"
 
 dependencies {
     implementation(projects.core.designSystem)
+    implementation(projects.core.common)
 }
