+++ feature/main/src/main/java/dev/enesky/feature/main/HomeScreen.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import dev.enesky.core.ui.LoadingWithTriangleDots
+import dev.enesky.core.ui.loading.LoadingWithTriangleDots
 import org.koin.androidx.compose.koinViewModel
 
 /**
